requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
databases==0.4.3
dataclasses==0.6
dnspython==2.1.0
fastapi==0.68.0
h11==0.12.0
idna==3.2
iniconfig==1.1.1
Jinja2==3.0.1
jwt==1.2.0
loguru==0.4.1
Mako==1.1.4
MarkupSafe==2.0.1
motor==2.5.0
numpy==1.21.1
packaging==21.0
pandas==1.3.1
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
pymongo==3.12.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
slugify==0.0.1
SQLAlchemy==1.3.24
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
wcwidth==0.2.5
websockets==9.1
win32-setctime==1.0.3
wincertstore==0.2
zipp==3.5.0
